feat(common) reorganize common folder with *LIST.hpp file to define component, system and script list
