+++ engine/src/RTypeEngine.hpp
@@ -48,7 +48,7 @@ class Engine
     void stop();
 
    private:
-    void addSystems();
+    // void addSystems();
     void loadGame();
 
     std::shared_ptr<IRuntime> _runtime = nullptr;         ///< Shared pointer to the runtime system.
