--- engine/src/common/cppScripts/helloworld.hpp
@@ -1,36 +1,0 @@
-/**********************************************************************************
- * Project: RType
- * Description: A GAME ENGINE THAT ROARS !
- * Author: Alexis BOITEL, Raphael MABILLE, Morgan LARGEOT, Maxence LARGEOT, Arthur DORIEL
- * Date, Location: 2024, Rennes
- **********************************************************************************/
-
-#ifndef HELLOWORLD_H
-#define HELLOWORLD_H
-
-#include "common/ICppScript.hpp"
-#include "gameContext/GameContext.hpp"
-
-namespace RType
-{
-
-class HelloWorldScript : public RType::ICppScript
-{
-   public:
-    void update(mobs::Registry &registry, GameContext &gameContext) override
-    {
-        if (gameContext._runtime->getKeyDown(KeyCode::H))
-        {
-            std::cout << "Hello, World! from " << _entity << std::endl;
-            gameContext._sceneManager._nextScene = "scenes1.json";
-        }
-    }
-    void setEntity(mobs::Entity entity) override { _entity = entity; }
-
-   private:
-    mobs::Entity _entity;
-};
-
-}  // namespace RType
-
-#endif  // HELLOWORLD_H
