+++ engine/src/gameContext/GameContext.hpp
@@ -12,8 +12,7 @@
 #include <sceneManager/SceneManager.hpp>
 
 #include "IRuntime/IRuntime.hpp"
-#include "common/components.hpp"
-#include "common/network/components/networkComp.hpp"
+#include "common/COMPONENTLIST.hpp"
 // std
 #include <chrono>
 #include <iostream>
